Add a file for additional/global NPM packages
"Sometimes it may also be required to supplement a packaging process
with additional NPM packages. For example, when building certain NPM
projects, some dependencies have to be installed globally.

A prominent example of such a workflow is a Grunt project. The grunt CLI
is typically installed globally, whereas its plugins are installed as
development dependencies." [0]

[0]: https://github.com/svanderburg/node2nix#adding-additionalglobal-npm-packages-to-a-packaging-process